restructure router
